[OPR-29]: Fixes the stability of the tests.
There was an issue where tests would pass in isolation, but fail when the entire suite was run. As expected - because the exclusion manager is a singleton, and it was keeping state between test runs.

At the moment, I've added a function which clears the exclusivity manager's state, but this must be public accessible for Swift 1.2.