[OPR-96]: Adds basic NSURLSessionTask op.
No unit tests for this, as it doesn't seem possible to subclass NSURLSession (and related clasess). Get an unrecognized selector error when trying to access the state. Issues seems to be that NSURLSession does some magic when it create NSURLSessionTasks (via dataTaskWithRequest etc), and the initializers for NSURLSession are no inherited.

To properly unit test this requires us to create a special task (i.e. a subclass) where we can assert that resume was called, and then immediately complete (to check that the operation finished). And to vend such a class (which is an NSURLSessionTask) requires constructing it either from a subclasses NSURLSession.

Regardless, it seems that accessing the state property fails.