[OPR-98]: Removes the alert condition on Authorize
This isn't really great at the moment, as it should really be there. However, there is a more deep rooted bug/issue with this which can use lock up. Essentially if you're nesting operations with conditions in groups, the group's conditions get added before the children, which can cause a dead lock if the child operations has the same condition.