[OPR-129]: Adds tests.
Adds coverage for more than one condition which is mutually exclusive with a mixture of Operation subclassed dependencies, and non-Operation subclass (NSBlockOperation) dependencies.

Essentially this test checks that the dependencies of conditions which are mutually exclusive are not executed concurrently, regardless of whether the dependent operation is a dumb NSOperation.