[OPR-171]: First step to fixing Reachability issue.
There are a few errors with the observers as of now.

Most common use case will be to attach an observer, which gets an immediate status response, and then detach if the status is "good", or have it return a Bool.

This could even be an internal element of the Reachability module which is solely exposed for ReachableOperation

Similarly, it needs to be smarter at scheduling & unscheduling. This should perhaps itself be Operations based. Because in the most common case (single obsever at once) it is probably quite inefficient. Instead it would be better to unschedule (if zero observers) after a slight delay. This gives an opportunity for another observer to prevent the un-schedule & re-schedule.