[OPR-118]: Updates Permissions example to show usage of AlertOperation
Because we execute a BlockOperation after the Alert *finishes* we can see that the AlertOperation is actually finishing. Additionally, if you set the log severity to .Verbose you can also see that the UIOperation also finishes.

It's actually a little tricky to test for this as a unit test - because it requires tapping the Okay button in the alert.