[OPR-197]: Fixes a bug where CloudKitOperation was getting multiple errors.
This is a little bit complicated, so I'm going to explain what is going on here.

1. CloudKitOperation is a GroupOperation subclass.
2. The child operations in this group are OPRCKOperation instances, which are themselves subclasses of ComposedOperation.
3. The composed operation is Apple's CKOperation
4. When composing NSOperation (not Operation) subclasses, they get put into a GroupOperation.
5. When ComposedOperation, which composes an NSOperation executes, it "produces" the GroupOperation.
6. When producing an operation, it gets added to the receiver's queue.

Therefore, all of this together means we have... A group operation, which has children which produces another GroupOperation. Therefore, the outer GroupOperation (the CloudKitOperation<T>) gets notified of when the produced operation (the GroupOperation which is wrapping T) finishes, but also, when it's own child (the OPRCKOperation / ComposedOperation) finishes. And, ComposedOperation also collects the errors of its composed operation.

This means that by default, GroupOperation will collect the error twice.

In summary, when GroupOperations, contain ComposedOperations which produce GroupOperations, we ultimately have two observers of when that GroupOperation finishes, which means we collect errors twice.

To get round this, I have modified GroupOperation to make it cleaner - we call a method which indicates the child did finish either errors. All it does is add the errors to the aggretate errors.

In CloudKitOperation<T> we override this, and call super, but only if the child is not a OPRCKOperation. Essenitally we ignore any errors coming from the composed operation, as we'll have already collected the same error.