Fixes bug accessing the PHPhotoLibrary (#282)
This was not getting trigged in unit tests, because the registrar property is lazy. And for all tests we replace the actual registrar object (in this case the PHPhotoLibrary) with a testable equivalent.

This means that the init() of PHPhotoLibrary was never called - and it turns out that it hits a run time exception.

This is exactly the sort of bug that Swift prevents, as if PHPhotoLibrary were written in Swift, the code would not have even compiled.