[OPR-293]: Fixes some issues with GroupOperation
One of the issues with adding the operations to the queue in the initializer, is that observers will not get notified.

So, we have to add the initial operations in `execute`. But, if we append to operations in addOperations, that means, we need to filter out any instances which have already been added.

For example, a subclass may well implement execute to first addOperation(myOperation); super.execute() which would cause a crash if we tried adding myOperation twice.