Will Cancel Observer (#293)
* [OPR-293]: Refactors Start & Cancellation observers

* [OPR-293]: Go back to original group cancellation policy

* [OPR-293]: Adds errors to will cancel observer

* [OPR-293]: Avoids overriding cancel

Instead adds a will cancel observer.

* [OPR-293]: Corrects mistake where the queue was not suspended

!!

* [OPR-293]: Makes cancel a final method

This prevents Operation subclasses from overriding cancel. Instead, they should add an observer as appropriate.

WillCancelObserver <- block is executed directly before the operation will change its cancellation state.

DidCancelObserver <- block is executed directly after the operation did change its cancellation state.

Typically, subclasses will want to observe WillCancel to update their own internal state. But consumers will want to observe DidCancel to respond to operations being cancelled.

* [OPR-293]: Adds to new operations to property

Also makes the queue private.

* [OPR-293]: Cancels all operations in the queue.

This will ensure that the finishing operation is also cancelled.

* [OPR-293]: Adds will/did cancel API

* [OPR-293]: Fixes some issues with GroupOperation

One of the issues with adding the operations to the queue in the initializer, is that observers will not get notified.

So, we have to add the initial operations in `execute`. But, if we append to operations in addOperations, that means, we need to filter out any instances which have already been added.

For example, a subclass may well implement execute to first addOperation(myOperation); super.execute() which would cause a crash if we tried adding myOperation twice.

* [OPR-293]: Renames will/didFinish

This is to be consistent with operationWill/operationDidCancel