Sample: GroupOperation with committed / uncommitted errors
Allows subclasses to commit changes to prior errors - for example, if
they’ve “handled” them.

Note: This is just a quick proof of concept. But it does fix this
particular issue with CloudKitOperation and the recovery handlers. (Not
a complete implementation, though - just erases the “handled” errors,
instead of committing them as handled.)