[OPR-390]: Fixes issues with ignored condition
This means that conditions which are ignored mean that the attached operation does not execute, but also does not finish with an error. It’s a nicer use case than GatedOperation. This works by checking the OperationResult for .Ignored and then only cancel() instead of cancelWithError().

Note that if the operation also has .Failed(errors) condition result, in addition to .Ignored then the operation will still not run but this time with errors (as expected).

Note that if the operation also has .Satisfied condition result, in addition to .Ignored then the operation will not run, but without errors.