Replace the assertionFailure if Condition.operation == nil with a log statement
Previously, if a ConditionOperation’s `operation` var was nil when it
executed, the following assertionFailure occurred:
`assertionFailure("ConditionOperation executed before operation set.”)`

However, this was not always correct, if the `operation` var was set,
but the operation cancelled/finished/went away prior to the
ConditionOperation being executed.

Since this case can happen, the `assertionFailure` has been replaced
with a more appropriate log statement.