Fix `Operation.cancelWithErrors()` thread-safety
The potential existed for `Operation.cancelWithErrors()` to result in
an assert due to an illegal state transition from .Pending =>
.Finishing while `cancelled == false`.

This was fixed by ensuring that the additional cancellation errors are
appended to the `_internalErrors` array at the same time that
`_cancelled` is set to `true` (i.e. within the same acquisition of the
lock).