Integrating recent changes from swift/2.2 branch into swift/2.3 branch (#424)
* Unschedules dispatch queue when stopping notifier (#360) (#410)

* Makes Operation's conditions property public accessible (#412)

* 3.1.1 (#411)

* Unschedules dispatch queue when stopping notifier (#360) (#410)

* Makes Operation's conditions property public accessible (#412)

* [3.1.1]: Adds CHANGELOG

* [3.1.1]: Updates version

* Fixes crash when accessing deallocated self (#416) (#417)

* fix: Crash when self was destroyed, issue #416

* ConditionOperation "executed before operation set" assertion fix (#420)

Fixes #415. 

* [OPR-415]: Replace the assertionFailure if Condition.operation == nil with a log statement

Previously, if a ConditionOperation’s `operation` var was nil when it executed, the following assertionFailure occurred:
`assertionFailure("ConditionOperation executed before operation set.”)`

However, this could also occur if the weak `operation` var was set, but the operation cancelled/finished/went away before the ConditionOperation was executed.

Thus the `assertionFailure` has been replaced with a log statement.

* [OPR-415]: Add a new StressTest case that covers some ConditionOperation race conditions and thread safety

* Fix for NetworkObserver Timer not properly handling cancellation (#344) (#421)

The value of `isCancelled` was being captured by the `dispatch_after` block at `init`, and thus the check always passed and the Timer's handler was always invoked (even if `Timer.cancel()` was called).
This was resolved by making `Timer` a reference type.

* Flickering / robustness fixes for NetworkObserver / NetworkIndicatorController (#422)

Previously, NetworkIndicatorController would not ensure that any existing Timer was cancelled when another observed (network) operation started prior to the Timer firing.

This resulted in potential flickering of the network activity indicator (or inaccurate state, where it disappeared too quickly) in some cases.

* Tests may randomly fail due to a bug in expectation fulfillment (#419)

Fixed a number of places where `XCTestExpectation.fulfill()` was not being called on the main thread:

* Main-thread expectation fulfillment in OperationTests.addCompletionBlockToTestOperation()

* Main-thread expectation fulfillment in BasicTests.swift

* Main-thread expectation fulfillment in StressTests.swift

* Increase StressTest timeout for CI

* Recode `test__completion_blocks` StressTest to use a dispatch_group instead of an expectation for every TestOperation.

* Main-thread expectation fulfillment in BlockConditionTests.swift

* Main-thread expectation fulfillment in CloudCapabilitiesTests.swift

* Main-thread expectation fulfillment in GroupOperationTests.swift

* Main-thread expectation fulfillment in NegatedConditionTests.swift

* Main-thread expectation fulfillment in RemoteNotificationConditionTests.swift

* Adjust StressTest to try to fix CI

`test__completion_blocks` used a `BlockOperation { }`, which (currently) defaults to dispatching the block to the main queue. Changing this to a `BlockOperation(block:)`, which uses the queue on which the BlockOperation is executed.

* Enhance internal testing `waitForOperation` methods to support an optional timeout parameter

* Increase timeouts for some StressTests to fix CI

* Fix `Operation.cancelWithErrors()` thread-safety (#423)

* Fix `Operation.cancelWithErrors()` thread-safety

The potential existed for `Operation.cancelWithErrors()` to result in an assert due to an illegal state transition from `.Pending` => `.Finishing` while `cancelled == false`.

This was fixed by ensuring that the additional cancellation errors are appended to the `_internalErrors` array at the same time that `_cancelled` is set to `true` (i.e. within the same acquisition of the lock).

* Added StressTest of `Operation.cancelWithErrors()`

* Rewrite `test__dependent_operations_always_run` to use a dispatch_group instead of an expectation for every BlockOperation

This speeds up the test, and resolves possible main-thread-stall issues running the test.