Mutual exclusive conditions (#448)
* [OPR-448]: Adds MutualExclusion

* [OPR-448]: Adds support for adding dependencies for mutual exclusion

* [OPR-448]: WIP on mutual exclusivity

* [OPR-448]: Finishes off test cases for mutual exclusivity

* [OPR-448]: Refactors add condition API to attach condition

I want to make the API consistent - but not choose consistency over intent. So, `add(dependency: myDependency)` makes sense - as it signals that the dependency is already managed by some other queue/system.

Where as for a Condition, we specifically _attach_ it to Procedure instance, and it does not (ideally should not) be managed outside the Procedure.

So, I want the concept of “attaching conditions to procedures” to be the language which is used to describe the behaviour.