Fix compilation with swift legacy version.
It was impossible to compile the project after settings the build
settings `SWIFT_VERSION=2.3`.

There was one ambiguity on the `dispatch_sync` function calls. I simply
added an intermediate struct to be able to call it explicitely.

There was one conflict with the Objective-C selector `atomic` when we
wanted to make `CKModifyRecordsOperation` adopting the
`CKModifyRecordsOperationType` protocol.
I renamed it as `atomicRecord`.