[OPR-544]: Fix test__cancel_block_procedure() failures
It was possible for the testing BlockProcedure to finish prior to the
`block.cancel()` call, which would result in some of the
BlockProcedures never cancelling (as they had finished first), and
a failing test.
Fixed by using a semaphore inside the BlockProcedure’s block.
Additionally, the test now waits for all BlockProcedures to cancel
*and* finish.