[OPR-566]: Refactors how conditions are evaluated
This fixes a existing issue where the ConditionEvaluation enum was basically just like ConditonResult, except that it could hold an array of errors.

This has been resolved by using an internal error type `ProcedureKitError.FailedConditions` to collect multiple errors from failed conditons, so that we can reduce a sequence of `ConditionResult`s into a single one.