Fix an occasional crash running the Location tests
It seems there may be an edge case in CoreLocation triggered when
creating (and releasing) lots of CLLocationManager instances (as
previously happened when each test, using its own
TestableLocationServicesRegistrar, created a “fake” CLLocationManager).

Change `TestableLocationServicesRegistrar`’s “fake” CLLocationManager
(which is only used to populate a parameter in the delegate callbacks)
to a static class variable.