Fixes to Location related procedures (#619)
* Capability.Location should set its registrar's delegate to nil on deinit.

CLLocationManager’s delegate is `unowned(unsafe)`, and we must ensure
it is set to nil before the object goes away. The only strong reference
to the delegate that Capability.Location sets is stored within
Capability.Location.

* LocationCapabilityTests tweaks

Explicitly set the test CLLocationManager’s `delegate` back to nil.
Explicitly wait on `capability.requestAuthorization` to fulfill an
expectation.

* ReverseGeocodeProcedureTests tweaks

* Fix an occasional crash running the Location tests

It seems there may be an edge case in CoreLocation triggered when
creating (and releasing) lots of CLLocationManager instances (as
previously happened when each test, using its own
TestableLocationServicesRegistrar, created a “fake” CLLocationManager).

Change `TestableLocationServicesRegistrar`’s “fake” CLLocationManager
(which is only used to populate a parameter in the delegate callbacks)
to a static class variable.