RetryIterator should not maintain a strong reference to the RetryProcedure
RetryIterator’s RetryFailureInfo (`info`) property is set right before
`addNextOperation()` is called, which calls `next()` on the
RetryIterator (which then uses the RetryFailureInfo).

Ultimately, the RetryFailureInfo contains strong references to the
RetryProcedure. And since the RetryProcedure owns the RetryIterator,
which owns the RetryFailureInfo… We end up with a reference cycle.

As an immediate quick fix, since the RetryFailureInfo is only used in
the call to `addNextOperation()`, set it back to `.none` after the call.