GroupProcedure: ensure that internal queue is emptied on deinit
Previously, a GroupProcedure that was created and released but never
added to a queue would leak the child operations on its internal
OperationQueue. To ensure that it is safe to release a GroupProcedure
in all circumstances, add a `deinit` handler that explicitly cancels
all operations on the queue, and also ensures that the queue is not
suspended (so the queue will handle the clean-up).