Procedure: Use an autoreleasepool in `start()` override
Replicate the `NSOperation.start()` autorelease behavior by pushing an
autoreleasepool before calling `main()`, and popping after `main()`
returns.
This is particularly important because we override the default
`start()` and `main()` behavior, and our `main()` calls the Procedure
subclass’s `execute()` (and we need to ensure that `execute()`, which
is running on a background thread, has an autoreleasepool).