Memory leaks fixes highlighted by Visual Memory Debugger (#624)
* TimeoutObserver should capture a weak reference to the procedure

* RetryIterator should not maintain a strong reference to the RetryProcedure

RetryIterator’s RetryFailureInfo (`info`) property is set right before
`addNextOperation()` is called, which calls `next()` on the
RetryIterator (which then uses the RetryFailureInfo).

Ultimately, the RetryFailureInfo contains strong references to the
RetryProcedure. And since the RetryProcedure owns the RetryIterator,
which owns the RetryFailureInfo… We end up with a reference cycle.

As an immediate quick fix, since the RetryFailureInfo is only used in
the call to `addNextOperation()`, set it back to `.none` after the call.

* Add test to catch reference cycles in RetryProcedure

(Fixed by prior commit.)

* GroupProcedure: ensure that internal queue is emptied on deinit

Previously, a GroupProcedure that was created and released but never
added to a queue would leak the child operations on its internal
OperationQueue. To ensure that it is safe to release a GroupProcedure
in all circumstances, add a `deinit` handler that explicitly cancels
all operations on the queue, and also ensures that the queue is not
suspended (so the queue will handle the clean-up).

* Revert "GroupProcedure: ensure that internal queue is emptied on deinit"

This reverts commit f2b7176fc8a96e841c8829f8d3b903633c50e883.

* GroupProcedure: ensure that internal queue is emptied on deinit

Previously, a GroupProcedure that was created and released but never
added to a queue would leak the child operations on its internal
OperationQueue. To ensure that it is safe to release a GroupProcedure
in all circumstances, add a `deinit` handler that explicitly cancels
all operations on the queue, and also ensures that the queue is not
suspended (so the queue will handle the clean-up).

* Procedure: Use an autoreleasepool in `start()` override

Replicate the `NSOperation.start()` autorelease behavior by pushing an
autoreleasepool before calling `main()`, and popping after `main()`
returns.
This is particularly important because we override the default
`start()` and `main()` behavior, and our `main()` calls the Procedure
subclass’s `execute()` (and we need to ensure that `execute()`, which
is running on a background thread, has an autoreleasepool).