ProcessProcedure thread-safety
`NSTask`/`Process` is listed as a Thread-Unsafe Class in the Foundation
Framework Thread Safety guide. As such, some major modifications to
ProcessProcedure were necessitated.

- `init(…)` now takes all the possible configuration parameters for the
Process, instead of a Process instance (the only required parameter is
`launchPath`)
- the ProcessProcedure is now responsible for creating the Process
object and managing it
- there is no public accessor for the Process object (it’s now an
internal implementation detail of ProcessProcedure)
- various additional methods / properties were added to expose Process
object functionality directly on the ProcessProcedure