Improve LogManager / Logger performance
- Provide custom `init()` variants for `_Logger` to ensure that only a
single LogManager lock acquisition is required.
- Adjust `LoggerProtocol` to not query the LogManager (global) state on
every call to `log(…)` (which was hurting performance even when logging
was disabled). Instead, each Procedure will pick up the current global
logging settings when it is instantiated, and can then be further
customized. To adjust the settings for all Procedures, simply set the
values on LogManager at program start.
- Switch LogManager to using a single PThreadMutex lock instead of a
Protector for each property for a performance win.