Procedure EventQueue (#668)
* Stress Test tweaks

* Improve NetworkObserver testing

* StressTestCase: set queue qualityOfService to .userInteractive

* Improve ProcedureKitTestCase

- Allow for setting `procedure` or `queue` to `nil` before test
tearDown.
- Set the FinishingProcedure’s name to something descriptive.

* Remove old MutualExclusivity tests

(The newer, robust MutualExclusivity tests cover these cases and more.)

* Procedure utilizes an internal EventQueue

The core of Procedure now utilizes an internal EventQueue - a serial
FIFO queue onto which all user “events” are dispatched - which results
in major thread-safety, performance, and ease-of-use wins, and a
reduction in framework overhead.

Numerous additional tests are included, edge-case bugs are fixed, and
code is simplified.

* Add CancellableResultProcedure / CancellableBlockProcedure

* Improve the BlockProcedureStressTests

* Clean-up Delay.swift

Now that Procedure utilizes an internal EventQueue, locks and
complexity can be reduced.

* Clean-up UserLocationProcedure, ReverseGeocodeProcedure

Now that it always safe to call `finish()` from a DidCancel observer,
complexity can be reduced.

* Improve LogManager / Logger performance

- Provide custom `init()` variants for `_Logger` to ensure that only a
single LogManager lock acquisition is required.
- Adjust `LoggerProtocol` to not query the LogManager (global) state on
every call to `log(…)` (which was hurting performance even when logging
was disabled). Instead, each Procedure will pick up the current global
logging settings when it is instantiated, and can then be further
customized. To adjust the settings for all Procedures, simply set the
values on LogManager at program start.
- Switch LogManager to using a single PThreadMutex lock instead of a
Protector for each property for a performance win.

* Clean-up NetworkDataProcedure

No longer needs a lock for `execute()`+DidCancelObserver.

* Clean-up NetworkDownloadProcedure

No longer needs a lock for `execute()`+DidCancelObserver.

* Clean-up NetworkUploadProcedure

No longer needs a lock for `execute()`+DidCancelObserver.

* Additional Stress Tests

* swiftlint tweaks

* ProcessProcedureTests tweak

* Improve TestableLocationManager, fix Location tests

* Minor clean-up and renaming

* Minor clean-up and renaming

* Minor clean-up

* Minor clean-up and renaming

* Renaming PendingEvents

* Minor renaming

* Minor clean-up and renaming

* Minor clean-ups / formatting

* Small tweaks and refactoring

* Tweaks and refactoring

* Rename futureExecute / futureFinish to pendingExecute / pendingFinish

* Replace the old Protector with MutexProtector (renamed to: Protector)

* Minor tweaks / reformatting

* Minor tweaks / renaming

* Update comments

* Update comments

* Additional tests

* Test tweak

* Minor tweak

Resolves a Leak-detector false-positive.