Improve AuthorizedFor mutual exclusivity
Previously, an `AuthorizedFor` condition ensured that the Procedure to
which it was added was mutually exclusive with any other Procedures
with the same `AuthorizedFor` condition (where the base
`CapabilityProtocol` was the same).

However, what is needed is for the `AuthorizeCapabilityProcedure`
(produced by the `AuthorizedFor` condition, and which can trigger a
prompt for the capability) to be mutually exclusive with other
`AuthorizeCapabilityProcedure`s with the same base CapabilityProtocol.

This change ensure that two Procedures do not try to prompt for the
same authorization/permission concurrently - but that once they have
acquired the authorization/permission, the Procedures may then execute
concurrently (unless some other MutuallyExclusive condition or
dependency prevents it).