[swift/3.1]: Removes a scenario which could cause an NSOperation to be waiting for itself
Bit of an odd situation where the GroupProcedure’s CanFinishGroup internal procedure holds the last strong reference to the group inside it’s execute method, which, causes the group to get deallocated when it goes out of scope, which in turn calls this removed line, which in turns calls `NSOperation.waitUntilFinished` on the CanFinishGroup. The documentation for this states: “An operation object must never call this method on itself and should avoid calling it on any operations submitted to the same operation queue as itself.” which is exactly what is happening.