Fixes Xcode 8.3 warnings for Swift 3.1 (#699)
* [swift/3.1]: Fixes warnings for Xcode 8.3

* [swift/3.1]: Fixes warning from Xcode 8.3 in tests

* [swift/3.1]: Fix for group errors issue.

* [swift/3.1]: Removing dedicated lint step - not needed, as part of build script anway.

* [swift/3.1]: Duh, tvOS is still 10.2 not 10.3.

* Group._procedureDidCancel should pass on the additional (cancellation) errors to the children (not all errors).

* [swift/3.1]: Removes a scenario which could cause an NSOperation to be waiting for itself

Bit of an odd situation where the GroupProcedure’s CanFinishGroup internal procedure holds the last strong reference to the group inside it’s execute method, which, causes the group to get deallocated when it goes out of scope, which in turn calls this removed line, which in turns calls `NSOperation.waitUntilFinished` on the CanFinishGroup. The documentation for this states: “An operation object must never call this method on itself and should avoid calling it on any operations submitted to the same operation queue as itself.” which is exactly what is happening.