Updates links to reference docs (#763)
Really not sure what the plan with the migration is, since I read all the old docs and found them beneficial but uncomfortable. 
Or maybe I have missed something glaringly obvious about what is going on. 
In any case I thought you may have not yet got the time to add the link.