[OPR-827]: Adds a test for result injection
Essentially, what is happening here, is that we are injecting the output of a procedure into the RepeatProcedure, as it’s associated payload conforms to InputProcedure.

The RepeatProcedure will then ensure that the input value is set on each subsequently created payload.

We can test this, my collecting the outputs of the payloads together via observers.